+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [Unreleased]
+### Added
+- `matches()` and `matchesSubstring()` shortcuts
+
+### Changed
+- More precise interpretation of `.`
+
 ## [0.0.2] - 2023-04-07
 ### Changed
 - API change